Add `isInstanceOf` to EnumOption.ts and Variable classes. Move FormatString related utils to `static` members on class
