+++ journeyapps-parser-schema/src/types/ObjectType.ts
@@ -18,7 +18,7 @@ export interface NotificationConfig {
 export class ObjectType extends Type {
   static readonly TYPE = 'object';
 
-  static isInstanceOf(type: Type): type is ObjectType {
+  static isInstanceOf(type: any): type is ObjectType {
     return type.name == ObjectType.TYPE;
   }
